python/pyproject.toml
[tool.poetry]
name = "agentsjson"
version = "0.1.0"
description = "Agents.json Python Implementation with Core and Integrations"
authors = ["Wildcard AI <info@wildcard.ai>"]
readme = "README.md"
packages = [
    {include = "agentsjson"}
]

[tool.poetry.dependencies]
python = "^3.8"
python-benedict = "^0.33.0"
pydantic = "^2.0.0"
typing-extensions = "^4.0.0"
pyyaml = "^6.0.0"
tweepy = "^4.0.0"
requests = "^2.0.0"
stripe = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 